Fix V3 app deletion test
We want to run a patched version of the cf-acceptance-tests to run
against capi-release 1.35.0. The behaviour of the API was changed,
but we are only upgrading capi-release and not cf-release, therefore
tests were breaking.

The branch including this fix will no longer be needed once we deploy
a version of cf-release that includes capi-release >= 1.35.0.