Increase timeout in loggregator firehose test.
We've seen this test failing intermittently. Increasing the timeout here
seems to resolve this. This also changes it to match the container
metrics test below which also uses 2*config.DefaultTimeoutDuration which
was increased in 1995884 to fix a similar issue.

In my testing, I often saw this failing against an AWS deployment,
whereas with this increased timeout, it didn't file after 20 rounds
using `ginkgo -untilItFails`. With some tracing added, I often saw it
taking a little over a minute for the messages to appear.